A lot of small changes
Added tower animations (but the towers are very badly implemented, so it's going to need a whole rewrite), tweaked the Enemy Spawner, the Fly enemy, the player and tower collision boxes, fixed a couple of minor bugs. Due to lack of time, a lot of code will need fixing and rewriting, as most are non-scalable and prone to error.