More bug fixes
Fixed the attack effect being tilted and just wrong in general, changed the Physics Overlap of the attack to a box one, but still needs adjustments since, for some reason, the attack sometimes just don´t register. Changed the Fly bullet, but it´s too small, going to change later.